api/microservices/gateway/requirements.txt
connexion ~= 2.13.0
connexion[swagger-ui]
flask-swagger-ui ~=3.36.0
grpc-interceptor ~= 0.11.0
grpcio-tools ~= 1.30
pymongo ~= 3.11.2
dnspython ~= 2.1.0
flask >= 2.0.3
python-dotenv >= 0.19.2
authlib >= 1.0
requests >= 2.27.1